BUGFIX: Glacier Badge may not boost SpcDef
BUGFIX: Glacier Badge may not boost Special Defense depending on the value of Special Attack.
Pryce's dialog ("That BADGE will raise the SPECIAL stats of POKéMON.") implies that Glacier Badge is intended to boost both Special Attack and Special Defense, but the Special Defense boost will not happen unless the unboosted Special Attack stat is 206–432, or 661 or above.